ui-jsx-fix: update ColorChip component
